fix: move navaccordion inside sidenav
