+++ packages/odyssey-react-mui/src/labs/index.ts
@@ -40,7 +40,7 @@ export {
   type GroupPickerProps,
 } from "./GroupPicker";
 
-export * from "./NavAccordion";
+export * from "./SideNav/NavAccordion";
 export * from "./SideNav";
 export * from "./TopNav";
 export * from "../ui-shell";
